+++ src/main/java/com/sparta/greeypeople/like/service/ReviewLikesService.java
@@ -1,5 +1,7 @@
-package com.sparta.greeypeople.like.review;
+package com.sparta.greeypeople.like.service;
 
+import com.sparta.greeypeople.like.entity.ReviewLikes;
+import com.sparta.greeypeople.like.repository.ReviewLikesRepository;
 import com.sparta.greeypeople.review.entity.Review;
 import com.sparta.greeypeople.review.repository.ReviewRepository;
 import com.sparta.greeypeople.user.entity.User;
